googleCloudFunctions/parse-save-grib-harmonie-v2/requirements.txt
functions-framework==3.*
numpy
requests
pytz
google-cloud-firestore
google-cloud-secret-manager
h5netcdf
scipy
datetime
python-dotenv